functions.sh: Fix grub-install to work with grub 2.04
The grub-efi version 2.04 expects that you will add any modules that
you might have run grub-mkimage with if using a Yocto Project build.

The arguments added to the efi version of the grub-install are the
same found in the grub-efi recipe.  This will eliminate the problem of
grub-efi printing a message that it is operating in "blind boot" mode
while trying to start the kernel.

The path for using the non-efi grub-install remains unchanged in the
overc installer because no additional modules are needed to start the
system.

[ Issue: LIN1019-2962 ]

(master rev: 3f240a12c5147ff09c1fe46a6fc7dea6251aeb3e

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>