cubeit-installer: Fixup warning on no network_prime
Issue: LIN1019-3409

The current WARNING will never be printed as the TODO if statement for
bridge networking will stop this from being hit. Under the assumption
the bridge networking is currently busted with no immediate plans to
fix it we move the WARNING to a location it will be displayed and we
also add and INFO and WARNING for network_vrf for consistency sake.

Upstream-Status: Backport

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>